refactor: response 타입들을 통합 관리하도록 수정
