+++ tsconfig.json
@@ -27,8 +27,6 @@
       "@libs/client": ["libs/client"],
       "@service/*": ["service/*"],
       "@utils/*": ["utils/*"],
-      "@types": ["types/index"],
-      "@types/*": ["types/*"],
       "@components/*": ["components/*"],
       "@hooks/*": ["hooks/*"],
       "@assets/*": ["public/assets/*"],
