+++ types/response.d.ts
@@ -1,0 +1,20 @@
+declare module 'apiResponse' {
+  interface ICategoryListAPIResponse {
+    id: number;
+    name: string;
+    url: string;
+    thumbnailId: string;
+  }
+
+  interface IPostDetailAPIResponse {
+    tag: {
+      tagName: string;
+    }[];
+    category: {
+      name: string;
+    } | null;
+    id: number;
+    title: string;
+    content: string;
+  }
+}
